requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
Flask==3.1.0
frozenlist==1.5.0
get-annotations==0.1.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-ollama==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.3.3
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.2
ollama==0.4.7
orjson==3.10.15
packaging==24.2
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.1
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yandex-cloud-ml-sdk==0.3.0
yandexcloud==0.331.0
yarl==1.18.3
zstandard==0.23.0
